pyproject.toml
[project]
name = "fluxmod-backend"
version = "0.1.0"
description = "Flask-based REST API for AutoMod rule management with Fluxer OAuth authentication."
readme = "README.md"
requires-python = ">=3.12.10"
dependencies = [
    "authlib>=1.6.9",
    "certifi>=2026.2.25",
    "colorama>=0.4.6",
    "fastapi>=0.135.1",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "fluxer-py>=0.4.1",
    "gunicorn>=25.1.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "pyright[nodejs]>=1.1.408",
    "ruff>=0.15.4",
]
